Fixan integer overflow bug that goes back at least as far as jai-imageio.  for  the fix in the java version see: https://github.com/faceless2/jpeg2000/commit/b20342956ac5d05bb97522ddeb41e9bed5abfa95
The code overflows an int in bp (bitplanes) is 30.  They have a complicated fix.  I just do the shifting in a long (which is essentially a free conversion on a 64 bit processor) and then cast it back to int when  the shifting is done.  (Also essentially free on most 64 bit processors.)